# web/packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Teggo B2B Commerce API
  version: 0.1.0
  description: >
    The single source of truth for the Teggo API contract. The TypeScript
    client used by both frontends is generated from this file (do not hand-edit
    the generated schema). Grown module-by-module as the backend lands.
servers:
  - url: http://localhost:8080
    description: Local Go API
...
components:
  securitySchemes:
    adminAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    storefrontAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    vendorAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    ConfigSetting:
      type: object
      required: [id, scope, key, value]
      properties:
        id: { type: integer, format: int64 }
        scope: { type: string, enum: [org, website, group, customer] }
        scope_id: { type: integer, format: int64, nullable: true }
        key: { type: string }
        value: {}
    ConfigSettingInput:
      type: object
      required: [scope, key, value]
      properties:
        scope: { type: string, enum: [org, website, group, customer] }
        scope_id: { type: integer, format: int64, nullable: true }
        key: { type: string }
        value: {}
    ListWrapperConfigSetting:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ConfigSetting' } }
    ConfigResolveResult:
      type: object
      required: [key, found]
      properties:
        key: { type: string }
        found: { type: boolean }
        scope: { type: string }
        value: {}
...
    PaymentConfig:
      type: object
      properties:
        gateway: { type: string }
        available: { type: array, items: { type: string }, description: 'Gateway adapters this deployment can use.' }
        configured_keys: { type: array, items: { type: string }, description: 'Names of stored credential keys — values are never returned.' }
    PaymentConfigUpdate:
      type: object
      required: [gateway]
      properties:
        gateway: { type: string }
        credentials: { type: object, additionalProperties: { type: string }, description: 'Replaces stored credentials (encrypted at rest). Omit to keep the existing ones; {} clears.' }
...
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /admin/settings:
    get:
      tags: [Settings]
      operationId: adminListSettings
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperConfigSetting' } } } }
    put:
      tags: [Settings]
      operationId: adminUpsertSetting
      security: [{ adminAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ConfigSettingInput' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ConfigSetting' } } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
  /admin/settings/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    delete:
      tags: [Settings]
      operationId: adminDeleteSetting
      security: [{ adminAuth: [] }]
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /admin/settings/resolve:
    get:
      tags: [Settings]
      operationId: adminResolveSetting
      summary: Preview the cascade-resolved value for a key at given scope ids.
      security: [{ adminAuth: [] }]
      parameters:
        - { name: key, in: query, required: true, schema: { type: string } }
        - { name: website_id, in: query, schema: { type: integer, format: int64 } }
        - { name: group_id, in: query, schema: { type: integer, format: int64 } }
        - { name: customer_id, in: query, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ConfigResolveResult' } } } }
  /storefront/settings/{key}:
    parameters:
      - { name: key, in: path, required: true, schema: { type: string } }
    get:
      tags: [Settings]
      operationId: storefrontResolveSetting
      summary: Resolve a config key for the current buyer (or org defaults if anonymous).
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ConfigResolveResult' } } } }
...
  /admin/settings/payments:
    get:
      tags: [Platform]
      operationId: getPaymentConfig
      summary: The org's payment gateway config — credential values are never returned.
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentConfig' } } } }
    put:
      tags: [Platform]
      operationId: updatePaymentConfig
      summary: Set the org's gateway and (optionally) replace its encrypted credentials.
      security: [{ adminAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PaymentConfigUpdate' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { gateway: { type: string } } } } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '503': { $ref: '#/components/responses/ErrorResponse' }
...